Move AstLineCounter to top level class
